--- src/index.css
@@ -1,137 +1,0 @@
-@font-face {
-  font-family: 'Inter';
-  src: url('./styles/fonts/Inter-SemiBold.ttf') format('truetype');
-  font-weight: 600;
-}
-@font-face {
-  font-family: 'Inter';
-  src: url('./styles/fonts/Inter-ExtraBold.ttf') format('truetype');
-  font-weight: 800;
-}
-@font-face {
-  font-family: 'Inter';
-  src: url('./styles/fonts/Inter-Bold.ttf') format('truetype');
-  font-weight: 700;
-}
-@font-face {
-  font-family: 'Inter';
-  src: url('./styles/fonts/Inter-Regular.ttf') format('truetype');
-  font-weight: 400;
-}
-
-:root {
-  padding: 0px;
-  margin: 0px;
-  box-sizing: border-box !important;
-  word-break: break-word;
-  --color-instance: #1e1e20;
-  --color-instance-popover-bg: #222222;
-  --Mail-Background: rgba(49, 51, 56, 1);
-  --new-message-text: black;
-  --bg-primary: rgba(31, 32, 35, 1);
-  --bg-2: #27282c;
-  --bg-3: rgba(0, 0, 0, 0.1);
-  --unread: #4297e2;
-  --danger: #b14646;
-  --apps-circle: #1f2023;
-  --green: #5eb049;
-}
-
-body {
-  margin: 0px;
-  overflow: hidden;
-}
-
-.image-container {
-  position: relative;
-}
-
-.image-container img {
-  position: absolute;
-  left: 0;
-  top: 0;
-  transition: opacity 0.5s ease; /* Optional: adds a fade effect */
-}
-
-.image-container .hover-image {
-  opacity: 0;
-}
-
-.image-container:hover .hover-image {
-  opacity: 1;
-}
-
-.image-container:hover .base-image {
-  opacity: 0.6;
-}
-
-::-webkit-scrollbar-track {
-  background-color: transparent;
-}
-::-webkit-scrollbar-track:hover {
-  background-color: transparent;
-}
-
-::-webkit-scrollbar {
-  width: 16px;
-  height: 10px;
-}
-
-::-webkit-scrollbar-thumb {
-  background-color: #444444;
-  border-radius: 8px;
-  background-clip: content-box;
-  border: 4px solid transparent;
-  transition: 0.3s background-color;
-}
-::-webkit-scrollbar-thumb:hover {
-  background-color: #363636;
-}
-
-@property --var1 {
-  syntax: '<color>';
-  inherits: true;
-  initial-value: transparent;
-}
-
-.scrollable-container {
-  transition: --var1 0.4s;
-}
-
-.scrollable-container:hover {
-  --var1: #444444;
-}
-
-.scrollable-container::-webkit-scrollbar-thumb {
-  background-color: var(--var1);
-  border-radius: 8px;
-  background-clip: content-box;
-  border: 4px solid transparent;
-  opacity: 0;
-}
-
-/* Mobile-specific scrollbar styles */
-@media only screen and (max-width: 600px) {
-  ::-webkit-scrollbar {
-    width: 8px; /* Narrower scrollbar width on mobile */
-    height: 6px; /* Narrower scrollbar height on mobile */
-  }
-
-  ::-webkit-scrollbar-thumb {
-    border-radius: 4px; /* Adjust the radius for a narrower thumb */
-    border: 2px solid transparent; /* Narrower thumb border */
-  }
-}
-
-.group-list::-webkit-scrollbar-thumb:hover {
-  background-color: whitesmoke;
-}
-
-html,
-body {
-  overscroll-behavior: none !important;
-}
-
-.swiper {
-  width: 100%;
-}
