+++ src/main.tsx
@@ -1,11 +1,10 @@
-import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
-import './index.css';
+import '../src/styles/index.css';
 import './messaging/messagesToBackground';
 import { MessageQueueProvider } from './MessageQueueContext.tsx';
 import { RecoilRoot } from 'recoil';
-import { ThemeProvider } from './styles/ThemeContext.tsx';
+import { ThemeProvider } from './components/Theme/ThemeContext.tsx';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <>
